Added format for the duration column and separated some functions
