--- resources/dummyData.js
@@ -1,303 +1,0 @@
-const dummyData = [{
-    "id": 1,
-    "status": "deleted",
-    "from": "937-986-8212",
-    "to": "581-687-4590",
-    "duration": "0:12:41"
-  }, {
-    "id": 2,
-    "status": "deleted",
-    "from": "163-932-2103",
-    "to": "244-360-9850",
-    "duration": "0:28:51"
-  }, {
-    "id": 3,
-    "status": "new",
-    "from": "569-245-1514",
-    "to": "160-275-1349",
-    "duration": "0:33:48"
-  }, {
-    "id": 4,
-    "status": "deleted",
-    "from": "707-212-0219",
-    "to": "562-444-9001",
-    "duration": "0:10:11"
-  }, {
-    "id": 5,
-    "status": "new",
-    "from": "514-857-8402",
-    "to": "744-498-4375",
-    "duration": "0:41:18"
-  }, {
-    "id": 6,
-    "status": "saved",
-    "from": "964-761-9717",
-    "to": "373-640-1577",
-    "duration": "0:44:08"
-  }, {
-    "id": 7,
-    "status": "deleted",
-    "from": "545-742-2820",
-    "to": "934-755-7653",
-    "duration": "0:19:05"
-  }, {
-    "id": 8,
-    "status": "deleted",
-    "from": "699-206-8762",
-    "to": "895-913-1506",
-    "duration": "0:04:21"
-  }, {
-    "id": 9,
-    "status": "deleted",
-    "from": "322-710-2678",
-    "to": "512-964-5001",
-    "duration": "0:52:32"
-  }, {
-    "id": 10,
-    "status": "deleted",
-    "from": "996-167-0788",
-    "to": "462-231-3706",
-    "duration": "0:48:55"
-  }, {
-    "id": 11,
-    "status": "new",
-    "from": "621-368-4459",
-    "to": "156-515-2167",
-    "duration": "0:40:15"
-  }, {
-    "id": 12,
-    "status": "deleted",
-    "from": "216-819-9405",
-    "to": "214-412-2197",
-    "duration": "0:02:19"
-  }, {
-    "id": 13,
-    "status": "new",
-    "from": "160-255-9187",
-    "to": "324-756-6793",
-    "duration": "0:08:06"
-  }, {
-    "id": 14,
-    "status": "saved",
-    "from": "786-779-8484",
-    "to": "123-449-2176",
-    "duration": "0:29:52"
-  }, {
-    "id": 15,
-    "status": "saved",
-    "from": "571-986-0013",
-    "to": "463-665-3911",
-    "duration": "0:41:22"
-  }, {
-    "id": 16,
-    "status": "new",
-    "from": "618-794-0304",
-    "to": "889-173-9706",
-    "duration": "0:09:18"
-  }, {
-    "id": 17,
-    "status": "saved",
-    "from": "615-328-3811",
-    "to": "398-192-9674",
-    "duration": "0:13:46"
-  }, {
-    "id": 18,
-    "status": "new",
-    "from": "456-120-3972",
-    "to": "144-422-4130",
-    "duration": "0:38:00"
-  }, {
-    "id": 19,
-    "status": "new",
-    "from": "146-532-6764",
-    "to": "492-215-4009",
-    "duration": "0:51:40"
-  }, {
-    "id": 20,
-    "status": "saved",
-    "from": "614-553-2798",
-    "to": "912-510-3615",
-    "duration": "0:03:21"
-  }, {
-    "id": 21,
-    "status": "new",
-    "from": "586-253-3002",
-    "to": "637-724-8955",
-    "duration": "0:18:42"
-  }, {
-    "id": 22,
-    "status": "new",
-    "from": "434-297-0297",
-    "to": "737-922-5767",
-    "duration": "0:18:12"
-  }, {
-    "id": 23,
-    "status": "saved",
-    "from": "380-228-8951",
-    "to": "926-532-6033",
-    "duration": "0:20:36"
-  }, {
-    "id": 24,
-    "status": "new",
-    "from": "102-392-2377",
-    "to": "414-892-9084",
-    "duration": "0:28:37"
-  }, {
-    "id": 25,
-    "status": "new",
-    "from": "989-843-9234",
-    "to": "450-558-1299",
-    "duration": "0:38:34"
-  }, {
-    "id": 26,
-    "status": "deleted",
-    "from": "404-125-2037",
-    "to": "326-159-5611",
-    "duration": "0:09:08"
-  }, {
-    "id": 27,
-    "status": "saved",
-    "from": "683-377-0021",
-    "to": "345-340-2395",
-    "duration": "0:46:27"
-  }, {
-    "id": 28,
-    "status": "deleted",
-    "from": "755-435-4832",
-    "to": "624-712-7408",
-    "duration": "0:06:15"
-  }, {
-    "id": 29,
-    "status": "new",
-    "from": "288-639-7881",
-    "to": "634-453-1906",
-    "duration": "0:49:46"
-  }, {
-    "id": 30,
-    "status": "saved",
-    "from": "515-363-4105",
-    "to": "440-348-3727",
-    "duration": "0:29:39"
-  }, {
-    "id": 31,
-    "status": "deleted",
-    "from": "810-920-7876",
-    "to": "774-780-2584",
-    "duration": "0:18:29"
-  }, {
-    "id": 32,
-    "status": "deleted",
-    "from": "934-516-8934",
-    "to": "791-765-1454",
-    "duration": "0:12:22"
-  }, {
-    "id": 33,
-    "status": "saved",
-    "from": "740-490-5792",
-    "to": "728-394-9351",
-    "duration": "0:16:49"
-  }, {
-    "id": 34,
-    "status": "deleted",
-    "from": "773-968-4744",
-    "to": "993-305-4667",
-    "duration": "0:52:54"
-  }, {
-    "id": 35,
-    "status": "saved",
-    "from": "416-119-0729",
-    "to": "835-881-8913",
-    "duration": "0:20:54"
-  }, {
-    "id": 36,
-    "status": "new",
-    "from": "831-575-9059",
-    "to": "692-785-5850",
-    "duration": "0:28:19"
-  }, {
-    "id": 37,
-    "status": "deleted",
-    "from": "294-966-8339",
-    "to": "912-563-1872",
-    "duration": "0:40:17"
-  }, {
-    "id": 38,
-    "status": "deleted",
-    "from": "187-731-2400",
-    "to": "944-366-0642",
-    "duration": "0:48:06"
-  }, {
-    "id": 39,
-    "status": "deleted",
-    "from": "627-335-5313",
-    "to": "967-694-5565",
-    "duration": "0:04:08"
-  }, {
-    "id": 40,
-    "status": "deleted",
-    "from": "127-601-6262",
-    "to": "796-968-1650",
-    "duration": "0:18:37"
-  }, {
-    "id": 41,
-    "status": "saved",
-    "from": "935-705-7277",
-    "to": "782-187-3112",
-    "duration": "0:31:22"
-  }, {
-    "id": 42,
-    "status": "deleted",
-    "from": "192-492-9150",
-    "to": "210-263-1681",
-    "duration": "0:35:55"
-  }, {
-    "id": 43,
-    "status": "saved",
-    "from": "118-462-6414",
-    "to": "140-219-0150",
-    "duration": "0:33:07"
-  }, {
-    "id": 44,
-    "status": "deleted",
-    "from": "217-701-2373",
-    "to": "868-929-0638",
-    "duration": "0:48:03"
-  }, {
-    "id": 45,
-    "status": "saved",
-    "from": "272-892-4195",
-    "to": "314-254-2607",
-    "duration": "0:34:25"
-  }, {
-    "id": 46,
-    "status": "saved",
-    "from": "506-730-5569",
-    "to": "618-660-5865",
-    "duration": "0:06:49"
-  }, {
-    "id": 47,
-    "status": "deleted",
-    "from": "831-481-1148",
-    "to": "911-529-3254",
-    "duration": "0:07:14"
-  }, {
-    "id": 48,
-    "status": "new",
-    "from": "812-512-3409",
-    "to": "883-251-2191",
-    "duration": "0:17:07"
-  }, {
-    "id": 49,
-    "status": "deleted",
-    "from": "833-978-7929",
-    "to": "907-676-4277",
-    "duration": "0:47:07"
-  }, {
-    "id": 50,
-    "status": "saved",
-    "from": "909-495-9049",
-    "to": "460-129-8723",
-    "duration": "0:06:07"
-  }];
-
-  export default dummyData;
