+++ src/components/VoicemailMessage.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import TableCell from '@material-ui/core/TableCell';
 import TableRow from '@material-ui/core/TableRow';
+import { formatDuration } from './../utils'; 
 
 const VoicemailMessage = ({ folder: status, to, from, length: duration }) => {
 
@@ -9,7 +10,7 @@ const VoicemailMessage = ({ folder: status, to, from, length: duration }) => {
               <TableCell align="center">{status}</TableCell>
               <TableCell align="right">{to}</TableCell>
               <TableCell align="right">{from}</TableCell>
-              <TableCell align="right">{duration}</TableCell>
+              <TableCell align="right">{formatDuration(duration)}</TableCell>
             </TableRow>
     );
 }
